requirements.txt
pyqlib>=0.9.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyyaml>=6.0
matplotlib>=3.5.0
tqdm>=4.62.0
akshare>=1.8.0
tushare>=1.2.0
yfinance>=0.2.0
requests>=2.28.0
streamlit>=1.20.0
plotly>=5.10.0
